% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{shapley}
\alias{shapley}
\title{Computes Shapley values for the explanatory variables in a Multiple Linear Regression model, using the coefficient of determination, R^2 and AIC to score coalitions.}
\usage{
shapley(Y, X)
}
\arguments{
\item{Y}{The response vector}

\item{X}{The design matrix of the linear regression. Should not include factor variables; convert to indicator variables first.}
}
\description{
Computes Shapley values for the explanatory variables in a Multiple Linear Regression model, using the coefficient of determination, R^2 and AIC to score coalitions.
}
\examples{
Y <- swiss[,6]
X <- as.matrix(swiss[,1:5])
shapley(Y,X)
}
